+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - APIManagerAPIAdapter keeps old parameters (See issue [#293](https://github.com/Axway-API-Management-Plus/apim-cli/issues/293))
 - Sample Dockerfile for apimcli
 - Environmentalize log level with environment variable LOG_LEVEL. Default LOG_LEVEL is info
+- Proxy Authentication is not working (See issue [#295](https://github.com/Axway-API-Management-Plus/apim-cli/issues/295))
 
 ## [1.12.1] 2022-07-06